Add redirect to login for new, edit and delete pages

+++ project.py
@@ -162,20 +162,20 @@ def gdisconnect():
 #JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
 def restaurantMenuJSON(restaurant_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return jsonify(MenuItems=[i.serialize for i in items])
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+  return jsonify(MenuItems=[i.serialize for i in items])
 
 
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
 def menuItemJSON(restaurant_id, menu_id):
-    Menu_Item = session.query(MenuItem).filter_by(id = menu_id).one()
-    return jsonify(Menu_Item = Menu_Item.serialize)
+  Menu_Item = session.query(MenuItem).filter_by(id = menu_id).one()
+  return jsonify(Menu_Item = Menu_Item.serialize)
 
 @app.route('/restaurant/JSON')
 def restaurantsJSON():
-    restaurants = session.query(Restaurant).all()
-    return jsonify(restaurants= [r.serialize for r in restaurants])
+  restaurants = session.query(Restaurant).all()
+  return jsonify(restaurants= [r.serialize for r in restaurants])
 
 
 #Show all restaurants
@@ -188,24 +188,28 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+  if 'username' not in login_session:
+    return redirect('/login')
   if request.method == 'POST':
-      newRestaurant = Restaurant(name = request.form['name'])
-      session.add(newRestaurant)
-      flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-      session.commit()
-      return redirect(url_for('showRestaurants'))
+    newRestaurant = Restaurant(name = request.form['name'])
+    session.add(newRestaurant)
+    flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+    session.commit()
+    return redirect(url_for('showRestaurants'))
   else:
-      return render_template('newRestaurant.html')
+    return render_template('newRestaurant.html')
 
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
-      if request.form['name']:
-        editedRestaurant.name = request.form['name']
-        flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-        return redirect(url_for('showRestaurants'))
+    if request.form['name']:
+      editedRestaurant.name = request.form['name']
+      flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
+      return redirect(url_for('showRestaurants'))
   else:
     return render_template('editRestaurant.html', restaurant = editedRestaurant)
 
@@ -213,6 +217,8 @@ def editRestaurant(restaurant_id):
 #Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
     session.delete(restaurantToDelete)
@@ -226,60 +232,65 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/')
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+  return render_template('menu.html', items = items, restaurant = restaurant)
      
 
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
+  if 'username' not in login_session:
+    return redirect('/login')
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
-      session.add(newItem)
-      session.commit()
-      flash('New Menu %s Item Successfully Created' % (newItem.name))
-      return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+    newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+    session.add(newItem)
+    session.commit()
+    flash('New Menu %s Item Successfully Created' % (newItem.name))
+    return redirect(url_for('showMenu', restaurant_id = restaurant_id))
   else:
-      return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+    return render_template('newmenuitem.html', restaurant_id = restaurant_id)
 
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
-
-    editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    if request.method == 'POST':
-        if request.form['name']:
-            editedItem.name = request.form['name']
-        if request.form['description']:
-            editedItem.description = request.form['description']
-        if request.form['price']:
-            editedItem.price = request.form['price']
-        if request.form['course']:
-            editedItem.course = request.form['course']
-        session.add(editedItem)
-        session.commit() 
-        flash('Menu Item Successfully Edited')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
+  if 'username' not in login_session:
+    return redirect('/login')
+  editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if request.method == 'POST':
+    if request.form['name']:
+        editedItem.name = request.form['name']
+    if request.form['description']:
+        editedItem.description = request.form['description']
+    if request.form['price']:
+        editedItem.price = request.form['price']
+    if request.form['course']:
+        editedItem.course = request.form['course']
+    session.add(editedItem)
+    session.commit() 
+    flash('Menu Item Successfully Edited')
+    return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+    return render_template('editmenuitem.html', restaurant_id = restaurant_id, menu_id = menu_id, item = editedItem)
 
 
 #Delete a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods = ['GET','POST'])
 def deleteMenuItem(restaurant_id,menu_id):
-    restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
-    itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
-    if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showMenu', restaurant_id = restaurant_id))
-    else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
+  if 'username' not in login_session:
+    return redirect('/login')
+  restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
+  if request.method == 'POST':
+    session.delete(itemToDelete)
+    session.commit()
+    flash('Menu Item Successfully Deleted')
+    return redirect(url_for('showMenu', restaurant_id = restaurant_id))
+  else:
+    return render_template('deleteMenuItem.html', item = itemToDelete)
